structure has been rebuilt
